refactor: get rid of old project structure and components
