+++ rs-class-components-app/src/app/components/artist/ArtistData.tsx
@@ -1,4 +1,4 @@
-import type { ArtistData } from '../../../types/artist-data';
+import type { ArtistData } from '../../types/artist-data';
 import type { ReactNode } from 'react';
 import { fetchArtistData } from '../../api/data-fetch';
