+++ rs-class-components-app/src/app/types/artist-data.ts
