+++ rs-class-components-app/src/app/types/component-properties.ts
