+++ rs-class-components-app/src/app/types/search-item.ts
