--- rs-class-components-app/src/context/ThemeContext.tsx
@@ -1,25 +1,0 @@
-import { createContext, type FC, type ReactNode, useState } from 'react';
-
-export type ThemeContextType = {
-  theme: string;
-  setTheme: (theme: string) => void;
-};
-
-export const ThemeContext = createContext<ThemeContextType | null>(null);
-
-export const ThemeContextProvider: FC<{ children: ReactNode }> = ({
-  children,
-}) => {
-  const [theme, setTheme] = useState<string>('light');
-
-  return (
-    <ThemeContext.Provider
-      value={{
-        theme,
-        setTheme,
-      }}
-    >
-      {children}
-    </ThemeContext.Provider>
-  );
-};
