--- rs-class-components-app/src/data/path-constants.ts
@@ -1,3 +1,0 @@
-export const MAIN = '/main';
-export const ABOUT = '/about';
-export const ANY = '/*';
