--- rs-class-components-app/src/features/artists/ArtistCard.css
@@ -1,15 +1,0 @@
-.card {
-  display: flex;
-  padding: 0.25rem 1rem;
-  width: 100%;
-  max-width: 600px;
-  gap: 0.5rem;
-  border: 1px solid #a8e1a4;
-  border-radius: 0.5rem;
-  cursor: pointer;
-  transition: all 0.3s linear;
-}
-
-.card:hover {
-  transform: scale(1.05);
-}
