--- rs-class-components-app/src/features/artists/ArtistDetailedCard.css
@@ -1,28 +1,0 @@
-.detailed-card {
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  gap: 1rem;
-  width: 40%;
-  text-align: left;
-  border: 1px solid #a8e1a4;
-  border-radius: 0.5rem;
-  height: fit-content;
-  padding: 1rem;
-}
-
-.artist-info {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.detailed-card a {
-  align-self: flex-end;
-  text-decoration: none;
-  color: inherit;
-}
-
-.detailed-card a:hover {
-  text-decoration: underline;
-}
