--- rs-class-components-app/src/features/error-boundary/ErrorBoundary.css
@@ -1,11 +1,0 @@
-.data-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: start;
-  gap: 0.5rem;
-  width: 100%;
-  max-width: 600px;
-  min-height: 85vh;
-  padding-bottom: 1rem;
-}
