--- rs-class-components-app/src/test-utils/server.ts
@@ -1,4 +1,0 @@
-import { setupServer } from 'msw/node';
-import { handlers } from './api-handlers.ts';
-
-export const server = setupServer(...handlers);
